# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Active Home Platform API
  version: 1.0.0
  summary: Configuration-driven housing & construction advertising marketplace (Uzbekistan) — v1 contract.
  description: "Enterprise OpenAPI 3.1 contract for **Active Home v1**, the implementation contract between\
    \ the\nweb frontend, backend (FastAPI modular monolith), AI coding agents, QA automation, future mobile\n\
    apps, and external integrations.\n\nThis document is generated from and fully aligned with the immutable\
    \ approved architecture:\nSRS v1.0, DDD Domain Model v1.0, Configuration & Metadata Framework v1.0,\
    \ SAD v1.0,\nDatabase Architecture & Logical Data Model v1.0, Physical Database Design v1.0, and the\n\
    API Architecture & Integration Specification v1.0.\n\n## Cross-cutting contract rules (see the API\
    \ Architecture Specification)\n- **Versioning:** URI major versioning under `/api/v1`. Within a major\
    \ version only additive,\n  backward-compatible change is allowed. Configuration changes (new categories,\
    \ forms, fields,\n  plans, roles, templates) are **never** contract changes — dynamic data flows through\
    \ stable\n  envelopes (`AttributeMap`, `LocalizedText`, `FormDefinition`), so the endpoint surface\
    \ is\n  invariant under the platform's normal mode of change.\n- **Authentication:** server-side sessions\
    \ backed by Redis (SAD §13). A session is established by an\n  auth flow and carried by a secure,\
...
  contact:
...
  license:
...
servers:
...
security:
- sessionCookie: []
...
paths:
  /auth/otp:
    post:
      tags:
      - Authentication
      operationId: requestOtp
      summary: Request a phone OTP
      description: 'Issues a single-use, expiring OTP via SMS (Eskiz) for registration, login, or recovery.

        **Rate limiting:** strict per-phone and per-IP throttle (`OtpThrottlePolicy`, NFR-SEC-004);

        repeated requests return 429 `OTP_THROTTLED`.

        '
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '202':
          description: OTP dispatched (no content reveals whether the phone is registered).
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/otp/verify:
    post:
      tags:
      - Authentication
      operationId: verifyOtp
      summary: Verify a phone OTP and establish a session
      description: 'Verifies the OTP and, on success, creates/authenticates the account and issues a first-party

        session (cookie + optional Bearer token). For REGISTRATION a new account is created; an existing

        verified phone is authenticated (FR-AUTH-001/004).

        '
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyRequest'
      responses:
        '200':
          description: Session established.
          headers:
            Set-Cookie:
              description: Sets the `ah_session` cookie for browser clients.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEstablished'
              examples:
                established:
                  $ref: '#/components/examples/SessionEstablishedExample'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/register/email:
    post:
      tags:
      - Authentication
      operationId: registerEmail
      summary: Register with email and password
      description: Creates an account pending email confirmation (FR-AUTH-002). Duplicate email is rejected
        409.
      security: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterEmailRequest'
      responses:
        '202':
          description: Account created; confirmation email sent.
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login/email:
    post:
      tags:
      - Authentication
      operationId: loginEmail
      summary: Log in with email and password
      description: Authenticates an email-credential account and establishes a session (FR-AUTH-004).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginEmailRequest'
      responses:
        '200':
          description: Session established.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEstablished'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login/google:
    post:
      tags:
      - Authentication
      operationId: loginGoogle
      summary: Federated sign-in with Google
      description: 'Exchanges a Google authorization code, then links (by verified email) or creates the
        account and

        issues a first-party session (FR-AUTH-003). Google is a sign-in provider only; no Google token
        is

        exposed to the client afterwards.

        '
      security:
      - googleOAuth:
        - openid
        - email
        - profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleSignInRequest'
      responses:
        '200':
          description: Session established (account linked or created).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEstablished'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /auth/logout:
    post:
      tags:
      - Authentication
      operationId: logout
      summary: Log out (terminate the current session)
      description: Revokes the current session immediately (FR-AUTH-005). The cookie is cleared.
      responses:
        '204':
          description: Session terminated.
          headers:
            Set-Cookie:
              description: Clears the `ah_session` cookie.
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/recovery:
    post:
      tags:
      - Authentication
      operationId: startRecovery
      summary: Start credential recovery
      description: Initiates recovery via a verified phone or email (FR-AUTH-006). Response does not reveal
        account existence.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryStartRequest'
      responses:
        '202':
          description: Recovery instructions dispatched if the identifier matches an account.
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /me:
    get:
      tags:
      - Users
      operationId: getMe
      summary: Get the authenticated account
      description: Returns the current account, privacy settings, preferences, owned profiles, and roles.
      parameters:
      - $ref: '#/components/parameters/Fields'
      responses:
        '200':
          description: The account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
      - Users
      operationId: updateMe
      summary: Update personal profile
      description: Updates display name and personal profile fields (FR-USER-001).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 120
                email:
                  type: string
                  format: email
              additionalProperties: false
      responses:
        '200':
          description: Updated account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /me/password:
    put:
      tags:
      - Users
      operationId: changePassword
      summary: Change password
      description: Changes the password for an email-credential account. Other sessions may be revoked
        server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204':
          description: Password changed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /me/preferences:
    put:
      tags:
      - Users
      operationId: updatePreferences
      summary: Update notification & privacy preferences
      description: Updates notification channel opt-ins and phone-reveal privacy (FR-USER-003; OTP always
        via SMS regardless).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                privacySettings:
                  $ref: '#/components/schemas/PrivacySettings'
                notificationPreferences:
                  $ref: '#/components/schemas/NotificationPreferences'
              additionalProperties: false
      responses:
        '200':
          description: Updated account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/account:
    delete:
      tags:
      - Users
      operationId: closeAccount
      summary: Close (anonymise) the account
      description: 'Closes the account. Modelled as **anonymisation, not deletion** (FR-USER-005, DM-06):
        personal

        identifiers are irreversibly replaced; the account id survives so listings, orders, and audit

        references remain valid.

        '
      responses:
        '204':
          description: Account closed and anonymised.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/sessions:
    get:
      tags:
      - Users
      operationId: listSessions
      summary: List active sessions
      description: Returns the account's active sessions (login history / device list).
      responses:
        '200':
          description: Sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/sessions/{sessionId}:
    delete:
      tags:
      - Users
      operationId: revokeSession
      summary: Revoke a session
      description: Revokes a specific session (remote logout).
      parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /me/sessions/switch-profile:
    post:
      tags:
      - Users
      operationId: switchActingProfile
      summary: Switch the acting profile for the current session
      description: 'Sets the acting context (personal or a specific owned business profile) on the current
        session

        (FR-USER-002). A session-level operation, not a re-authentication.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actingProfileId:
                  type:
                  - string
                  - 'null'
                  format: uuid
                  description: null = personal context.
              additionalProperties: false
      responses:
        '200':
          description: Updated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /admin/users:
    get:
      tags:
      - Administration
      operationId: adminListUsers
      summary: List users (admin)
      description: Administrative user search/listing. Requires user-management permission.
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - CLOSED
      - name: query
        in: query
        schema:
          type: string
          description: Phone/email/name partial.
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of users.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/CursorPage'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/UserAdminView'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ah_session
      description: 'First-party, server-side session (Redis-backed, SAD §13). Issued by any auth flow
        and sent as a

        secure, HttpOnly, SameSite=Lax cookie. This is the default scheme for browser clients.

        '
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 'The same first-party session token presented as `Authorization: Bearer <token>` for
        native and

        mobile clients that cannot use cookies. The token is an **opaque server-side session reference**,

        not a JWT. Revoking the session (logout) invalidates it immediately.

        '
    googleOAuth:
      type: oauth2
      description: 'Google is a federated **sign-in** identity provider only (authorization-code flow).
        After Google

        authenticates the user, Active Home links or creates the account and issues its own first-party

        session (`sessionCookie`/`sessionBearer`). Active Home does not expose an OAuth2 authorization

        server of its own.

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect sign-in
...